backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
qdrant-client==1.7.0
openai==1.3.7
anthropic==0.7.8
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2
python-dateutil==2.8.2
boto3==1.29.7
redis==5.0.1
celery==5.3.4
pydub==0.25.1
httpx==0.25.2
sse-starlette==1.8.2

